Use stack_entry to represent method return
* use stack_entry_t so method can return different types
* split functions into classfile.c and constant-pool.c
